SDK - Moved python op pipeline compilation test to bridge tests
